+++ src/commands/preact.js
@@ -1,15 +1,12 @@
-#!/usr/bin/env node
-
 import {cyan as opt, green as cmd, red, yellow as req} from 'chalk'
 import parseArgs from 'minimist'
 
-import pkg from '../../package.json'
 import {CONFIG_FILE_NAME} from '../constants'
 import {ConfigValidationError, UserError} from '../errors'
 
 const COMMAND_MODULES = {
-  build: 'build-inferno',
-  run: 'serve-inferno',
+  build: 'build-preact',
+  run: 'serve-preact',
 }
 
 function handleError(error) {
@@ -28,35 +25,25 @@ function handleError(error) {
   process.exit(1)
 }
 
-let args = parseArgs(process.argv.slice(2), {
+let args = parseArgs(process.argv.slice(3), {
   alias: {
     c: 'config',
-    h: 'help',
     p: 'plugins',
-    v: 'version',
-  },
-  boolean: ['help', 'version'],
+  }
 })
 
 let command = args._[0]
 
-if (args.version || /^v(ersion)?$/.test(command)) {
-  console.log(`v${pkg.version}`)
-  process.exit(0)
-}
-
-if (args.help || !command || /^h(elp)?$/.test(command)) {
-  console.log(`Usage: ${cmd('inferno')} ${req('(run|build)')} ${opt('[options]')}
+if (!command || /^h(elp)?$/.test(command)) {
+  console.log(`Usage: ${cmd('nwb preact')} ${req('(run|build)')} ${opt('[options]')}
 
 Options:
   ${opt('-c, --config')}   config file to use ${opt(`[default: ${CONFIG_FILE_NAME}]`)}
-  ${opt('-h, --help')}     display this help message
   ${opt('-p, --plugins')}  a comma-separated list of nwb plugins to use
-  ${opt('-v, --version')}  print nwb's version
 
 Commands:
-  ${cmd('inferno run')} ${req('<entry>')} ${opt('[options]')}
-    Serve an Inferno app or component module.
+  ${cmd('nwb preact run')} ${req('<entry>')} ${opt('[options]')}
+    Serve a Preact app or component module.
 
     Arguments:
       ${req('entry')}          entry point for the app, or a component module
@@ -70,10 +57,10 @@ Commands:
       ${opt('--no-polyfill')}  disable inclusion of default polyfills
       ${opt('--port')}         port to run the dev server on ${opt('[default: 3000]')}
       ${opt('--reload')}       auto reload the page if hot reloading fails
-      ${opt('--title')}        contents for <title> ${opt('[default: Inferno App]')}
+      ${opt('--title')}        contents for <title> ${opt('[default: Preact App]')}
 
-  ${cmd('inferno build')} ${req('<entry>')} ${opt('[dist_dir] [options]')}
-    Create a static build for an Inferno app.
+  ${cmd('nwb preact build')} ${req('<entry>')} ${opt('[dist_dir] [options]')}
+    Create a static build for a Preact app.
 
     Arguments:
       ${req('entry')}       entry point for the app
@@ -83,18 +70,18 @@ Commands:
       ${opt('--force')}        don't shim rendering, use the given entry module directly
       ${opt('--mount-id')}     id for the <div> the app will render into ${opt('[default: app]')}
       ${opt('--no-polyfill')}  disable bundling of default polyfills
-      ${opt('--title')}        contents for <title> ${opt('[default: Inferno App]')}
+      ${opt('--title')}        contents for <title> ${opt('[default: Preact App]')}
       ${opt('--vendor')}       create a 'vendor' bundle for node_modules/ modules
 `)
-  process.exit(args.help || command ? 0 : 1)
+  process.exit(command ? 0 : 1)
 }
 
 if (!COMMAND_MODULES.hasOwnProperty(command)) {
-  console.error(`${red('Unknown command:')} ${req(command)}`)
+  console.error(`${red('Unknown preact command:')} ${req(command)}`)
   process.exit(1)
 }
 
-let commandModule = require(`../commands/${COMMAND_MODULES[command]}`)
+let commandModule = require(`./${COMMAND_MODULES[command]}`)
 
 try {
   commandModule(args, err => {